test: consolidate tests
Unify how we test in parse adapters and reintroduce
fake browser testing (using JSDOM).
