+++ packages/apidom-parser-adapter-asyncapi-json-2/.mocharc.js
@@ -3,5 +3,5 @@
 module.exports = {
   recursive: true,
   spec: 'test/**/*.ts',
-  require: ['test/mocha-bootstrap.js'],
+  file: ['test/mocha-bootstrap.js'],
 };
