+++ packages/apidom-parser-adapter-asyncapi-json-2/test/adapter.ts
@@ -1,7 +1,7 @@
 import fs from 'fs';
 import path from 'path';
-import { assert } from 'chai';
-import { isParseResultElement } from '@swagger-api/apidom-core';
+import { assert, expect } from 'chai';
+import { isParseResultElement, sexprs } from '@swagger-api/apidom-core';
 import { isAsyncApi2Element } from '@swagger-api/apidom-ns-asyncapi-2';
 
 import * as adapter from '../src/adapter';
@@ -18,6 +18,7 @@ describe('adapter', function () {
 
     assert.isTrue(isParseResultElement(parseResult));
     assert.isTrue(isAsyncApi2Element(parseResult.api));
+    expect(sexprs(parseResult)).toMatchSnapshot();
   });
 
   context('given zero byte empty file', function () {
