+++ packages/apidom-parser-adapter-openapi-json-3-1/test/adapter.ts
@@ -1,7 +1,7 @@
 import fs from 'fs';
 import path from 'path';
-import { assert } from 'chai';
-import { isParseResultElement } from '@swagger-api/apidom-core';
+import { assert, expect } from 'chai';
+import { isParseResultElement, sexprs } from '@swagger-api/apidom-core';
 import { isOpenApi3_1Element } from '@swagger-api/apidom-ns-openapi-3-1';
 
 import * as adapter from '../src/adapter';
@@ -18,6 +18,7 @@ describe('adapter', function () {
 
     assert.isTrue(isParseResultElement(parseResult));
     assert.isTrue(isOpenApi3_1Element(parseResult.api));
+    expect(sexprs(parseResult)).toMatchSnapshot();
   });
 
   context('given zero byte empty file', function () {
