+++ packages/apidom-parser-adapter-yaml-1-2/test/adapter-browser.ts
@@ -1,10 +1,33 @@
-import { assert } from 'chai';
+import fs from 'fs';
+import path from 'path';
+import { assert, expect } from 'chai';
 import dedent from 'dedent';
-import { SourceMapElement } from '@swagger-api/apidom-core';
+import {
+  isObjectElement,
+  isParseResultElement,
+  sexprs,
+  SourceMapElement,
+} from '@swagger-api/apidom-core';
 
-import * as adapter from '../src/adapter-node';
+import * as adapter from '../src/adapter-browser';
+
+const spec = fs.readFileSync(path.join(__dirname, 'fixtures', 'sample-data.yaml')).toString();
 
 describe('adapter-browser', function () {
+  it('should not detect proper media type', function () {
+    assert.isFalse(adapter.detect(spec));
+  });
+
+  it('should parse', async function () {
+    const parseResult = await adapter.parse(spec, {
+      sourceMap: true,
+    });
+
+    assert.isTrue(isParseResultElement(parseResult));
+    assert.isTrue(isObjectElement(parseResult.result));
+    expect(sexprs(parseResult)).toMatchSnapshot();
+  });
+
   context('given zero byte empty file', function () {
     specify('should return empty parse result', async function () {
       const parseResult = await adapter.parse('', { sourceMap: true });
