+++ packages/apidom-parser-adapter-yaml-1-2/test/mocha-bootstrap.js
@@ -3,6 +3,19 @@
 require('@babel/register')({ extensions: ['.js', '.ts'], rootMode: 'upward' });
 
 /**
+ * Configure snapshot testing.
+ */
+const chai = require('chai');
+const { jestSnapshotPlugin, addSerializer } = require('mocha-chai-jest-snapshot');
+
+const jestApiDOMSerializer = require('../../../scripts/jest-serializer-apidom');
+const jestStringSerializer = require('../../../scripts/jest-serializer-string');
+
+chai.use(jestSnapshotPlugin());
+addSerializer(jestApiDOMSerializer);
+addSerializer(jestStringSerializer);
+
+/**
  * JSDOM setup.
  */
 require('jsdom-global')();
